+++ nodes/parachain/Cargo.toml
@@ -122,6 +122,4 @@ std = [
 	"xcm/std",
 ]
 on-chain-release-build = []
-async-backing = [
-	"politest-runtime/async-backing"
-]
+async-backing = [ "politest-runtime/async-backing" ]
